+++ include/asio_sodium/detail/message_header.hpp
@@ -69,7 +69,7 @@ namespace detail {
     data_span() noexcept { return view_.last<data_size>(); }
 
     auto
-    buffer() noexcept { return asio::buffer(&view_[0], view_.size()); }
+    buffer() noexcept { return asio::buffer(&view_[0], static_cast<std::size_t>(view_.size())); }
 
     constexpr nonce_span
     data_nonce_field() noexcept {
@@ -144,7 +144,7 @@ namespace detail {
         crypto_box_open_easy(
           &data_span[0]
         , &full_span[0]
-        , full_span.size()
+        , static_cast<std::size_t>(full_span.size())
         , &nonce[0]
         , &public_key[0]
         , &private_key[0]
@@ -160,13 +160,13 @@ namespace detail {
     void
     generate_data_nonce() noexcept {
       auto data_nonce = view_.data_nonce_field();
-      randombytes_buf(&data_nonce[0], data_nonce.size());
+      randombytes_buf(&data_nonce[0], static_cast<std::size_t>(data_nonce.size()));
     }
 
     void
     generate_followup_nonce() noexcept {
       auto followup_nonce = view_.followup_nonce_field();
-      randombytes_buf(&followup_nonce[0], followup_nonce.size());
+      randombytes_buf(&followup_nonce[0], static_cast<std::size_t>(followup_nonce.size()));
     }
 
     void
@@ -241,7 +241,7 @@ namespace detail {
         crypto_box_easy(
           &full_span[0]
         , &data_span[0]
-        , data_span.size()
+        , static_cast<std::size_t>(data_span.size())
         , &nonce[0]
         , &public_key[0]
         , &private_key[0]
